Add QGS and MPA registering in the virt-handler
The Quote Generation Service (QGS) and the MPA registering service are
required to allow a confidential guest to generate a blob. The MPA
registering runs only once and registers the hw to a remote server. The
QGS talks with QEMU when the guest requires to generate a blob. This
communication happens by using a unix socket. Add the QGS and the MPA
registering components as new containers in the virt-handler pod. Then,
use a socket device plugin to mount the unix socket in the
virt-launcher. Create a new image named qgs-launcher-image that contains
the tdx-qgs and the sgx-mpa components. This image is used for both the
qgs-launcher container and the mpa-registering container.

Signed-off-by: Matias Ezequiel Vara Larsen <mvaralar@redhat.com>